ref(service): Move SendTraced into backend extensions module
