+++ objectstore-service/src/backend/mod.rs
@@ -15,10 +15,10 @@ pub mod bigtable;
 pub mod changelog;
 pub mod common;
 pub mod counting;
+mod extensions;
 pub mod gcs;
 pub mod in_memory;
 pub mod local_fs;
-mod response;
 pub mod s3_compatible;
 pub mod tiered;
